requirements.txt
streamlit
Pyrebase4
PyPDF2
google-generativeai
streamlit-calendar